+++ test/src/analyzers/lint_analyzer/rules/rules_list/no_magic_number/examples/enum_example.dart
@@ -1,0 +1,17 @@
+enum ExampleMagicNumbers {
+  second(2),
+  third(3);
+
+  final int value;
+
+  const ExampleMagicNumbers(this.value);
+}
+
+enum ExampleNamedMagicNumbers {
+  second(value: 2),
+  third(value: 3);
+
+  final int value;
+
+  const ExampleNamedMagicNumbers({required this.value});
+}