+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * feat: add `ignored-types` option to [`avoid-late-keyword`](https://dcm.dev/docs/individuals/rules/common/avoid-late-keyword).
 * fix: support tear-off methods for `check-unnecessary-nullable`.
 * fix: correctly handle empty comment sentence for [`format-comment`](https://dcm.dev/docs/individuals/rules/common/format-comment).
+* feat: support type=lint suppression.
 
 ## 5.4.0